requirements312.txt
python-telegram-bot[rate-limiter]
openai
tiktoken
PyYAML
pymongo
python-dotenv
